move: types/api/models/prefectureがprefecturesになっていたのでprefectureに移動させた
